fix #809 and minor bug in pool
