+++ capture/sock.go
@@ -1,9 +1,10 @@
+// +build linux
+
 package capture
 
 import (
 	"fmt"
 	"net"
-	"sync"
 	"time"
 	"unsafe"
 
@@ -30,18 +31,6 @@ const (
 )
 
 var tpacket2hdrlen = tpAlign(int(unsafe.Sizeof(unix.Tpacket2Hdr{})))
-
-// SockRaw is a linux M'maped af_packet socket
-type SockRaw struct {
-	mu          sync.Mutex
-	fd          int
-	ifindex     int
-	snaplen     int
-	pollTimeout uintptr
-	frame       uint32 // current frame
-	buf         []byte // points to the memory space of the ring buffer shared with the kernel.
-	loopIndex   int32  // this field must filled to avoid reading packet twice on a loopback device
-}
 
 // NewSockRaw returns new M'maped sock_raw on packet version 2.
 func NewSockRaw(ifi net.Interface) (*SockRaw, error) {
